ConnectorStackAttributesProvider refactor.
